working on Deleted Thread by authorized users

+++ app/Http/Controllers/ThreadsController.php
@@ -112,6 +112,14 @@ class ThreadsController extends Controller
      */
     public function destroy($channel, Thread $thread)
     {
+        if ( $thread->user_id != auth()->id() )
+        {
+            if ( request()->wantsJson() )
+            {
+                return response([ 'status' => 'Permission Denied', 403 ]);
+            }
+            return redirect('login');
+        }
 
         $thread->delete();
         if ( request()->wantsJson() )
